Split impl Serializer for Record
